I made an improvement to the SampleFile2.
This led to a hunt for NULL-pointer issues, that led me to making "Report" a Command, rather than a function call. Report left the AI engine in a useful, though invalid, state. So, either all of the functions that modify the AI state had to guard against this, or I had to change how Report worked. I tried the guard first, and it made me realize that standard library functions should only throw TypedOperationExceptions, so that they produce higher quality error messages (FatalException doesn't build a stack trace: FatalException is for player use only, and the player is responsible for the quality of the error message). Then I realized that "call Command('Left') call Report()" was valid, and I didn't like what it would do. So, Report became a Command.